LOGIN: updating login count in login file
