+++ src/components/login/login.types.ts
@@ -15,6 +15,7 @@ export interface LoginFormState {
         password: string;
     };
     loading: boolean;
+    loginCount: number
 }
 
 export interface LoginProps extends RouteComponentProps<{}> { }
